+++ Text/CustomNER/custom_ner/function.json
@@ -1,5 +1,5 @@
 {
-  "scriptFile": "__init__.py",
+  "scriptFile": "main.py",
   "bindings": [
     {
       "authLevel": "function",
